+++ src/Handlers/CrawlRequestFailed.php
@@ -7,6 +7,7 @@ use GuzzleHttp\Exception\ConnectException;
 use GuzzleHttp\Exception\RequestException;
 use GuzzleHttp\Psr7\Request;
 use Spatie\Crawler\Crawler;
+use Spatie\Crawler\TransferStatistics;
 
 class CrawlRequestFailed
 {
@@ -24,8 +25,13 @@ class CrawlRequestFailed
             $exception = new RequestException($exception->getMessage(), $request, previous: $exception);
         }
 
+        $transferStats = $this->crawler->getTransferStats($crawlUrl->url);
+        $transferStatistics = $transferStats !== null
+            ? TransferStatistics::fromTransferStats($transferStats)
+            : null;
+
         $this->crawler->recordFailed();
-        $this->crawler->getCrawlObservers()->crawlFailed($crawlUrl, $exception, $this->crawler->getCrawlProgress());
+        $this->crawler->getCrawlObservers()->crawlFailed($crawlUrl, $exception, $this->crawler->getCrawlProgress(), $transferStatistics);
 
         $this->crawler->applyDelay();
     }